requirements.txt
accelerate==0.24.1
asgiref==3.7.2
autopep8==2.0.4
certifi==2023.11.17
charset-normalizer==3.3.2
diffusers==0.23.1
Django==4.2.7
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.8.0
invisible-watermark==0.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
psutil==5.9.6
pycodestyle==2.11.1
PyWavelets==1.5.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
sqlparse==0.4.4
sympy==1.12
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.1
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
urllib3==2.1.0
zipp==3.17.0
